pyproject.toml
[project]
name = "openai-agents-sdk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm,viz]>=0.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
